+++ tilebox-workflows/tests/tasks_data.py
@@ -14,24 +14,31 @@ from hypothesis.strategies import (
     dictionaries,
     floats,
     integers,
+    just,
     lists,
     none,
     one_of,
     recursive,
     sampled_from,
     text,
+    timedeltas,
     uuids,
 )
 
+from tilebox.datasets.query.id_interval import IDInterval
+from tilebox.datasets.query.time_interval import TimeInterval
 from tilebox.workflows.data import (
     AutomationPrototype,
     Cluster,
     ComputedTask,
     CronTrigger,
+    ExecutionStats,
     Idling,
     Job,
     JobState,
     ProgressIndicator,
+    QueryFilters,
+    SingleTaskSubmission,
     StorageEventTrigger,
     StorageLocation,
     StorageType,
@@ -39,7 +46,8 @@ from tilebox.workflows.data import (
     TaskIdentifier,
     TaskLease,
     TaskState,
-    TaskSubmission,
+    TaskSubmissionGroup,
+    TaskSubmissions,
 )
 
 
@@ -118,47 +126,121 @@ def task_inputs(draw: DrawFn) -> bytes:
 
 
 @composite
-def jobs(draw: DrawFn, canceled: bool | None = None) -> Job:
+def execution_stats(draw: DrawFn) -> ExecutionStats:
+    """A hypothesis strategy for generating random execution_stats"""
+    first_task_started_at = draw(
+        one_of(none(), datetimes(min_value=datetime(1990, 1, 1), max_value=datetime(2024, 1, 1)))
+    )
+    last_task_stopped_at = None
+    if first_task_started_at is not None:
+        # only if we have a first task started at, we can have a last task stopped at, but it can also be none still
+        last_task_stopped_at = draw(
+            one_of(none(), datetimes(min_value=first_task_started_at, max_value=datetime(2024, 1, 1)))
+        )
+        first_task_started_at = first_task_started_at.astimezone(timezone.utc)
+    if last_task_stopped_at is not None:
+        last_task_stopped_at = last_task_stopped_at.astimezone(timezone.utc)
+
+    elapsed_time = draw(timedeltas(min_value=timedelta(seconds=0), max_value=timedelta(days=1)))
+    compute_time = draw(timedeltas(min_value=timedelta(seconds=0), max_value=timedelta(days=1)))
+    parallelism = draw(floats(min_value=0, max_value=100))
+
+    tasks_by_state = {}
+    for state in TaskState:
+        n = draw(integers(min_value=0, max_value=1000))
+        if n > 0:
+            tasks_by_state[state] = draw(integers(min_value=0, max_value=1000))
+
+    if len(tasks_by_state) == 0:
+        tasks_by_state[TaskState.QUEUED] = 1  # at least one task must be in the stats
+
+    return ExecutionStats(
+        first_task_started_at,
+        last_task_stopped_at,
+        compute_time,
+        elapsed_time,
+        parallelism,
+        sum(tasks_by_state.values()),
+        tasks_by_state,
+    )
+
+
+@composite
+def jobs(draw: DrawFn) -> Job:
     """A hypothesis strategy for generating random jobs"""
     job_id = draw(uuids(version=4))
     name = draw(alphanumerical_text())
     trace_parent = draw(alphanumerical_text())
     state = draw(sampled_from(JobState))
-    submitted_at = draw(datetimes(min_value=datetime(1990, 1, 1), max_value=datetime(2024, 1, 1)))
-    started_at = draw(
-        one_of(
-            none(),
-            datetimes(min_value=submitted_at, max_value=datetime(2025, 1, 1)),
-        )
+    submitted_at = draw(
+        datetimes(min_value=datetime(1990, 1, 1), max_value=datetime(2024, 1, 1), timezones=just(timezone.utc))
     )
-    if canceled is None:
-        canceled = draw(booleans())
-
     progress = draw(lists(progress_indicators(), min_size=0, max_size=3))
+    stats = draw(execution_stats())
 
     return Job(
         job_id,
         name,
         trace_parent,
         state,
-        submitted_at.astimezone(timezone.utc),
-        started_at.astimezone(timezone.utc) if started_at else None,
-        canceled,
+        submitted_at,
         progress,
+        stats,
     )
 
 
 @composite
-def task_submissions(draw: DrawFn) -> TaskSubmission:
+def task_submission_groups(draw: DrawFn) -> TaskSubmissionGroup:
+    """A hypothesis strategy for generating random task_submission_groups"""
+    dependencies_on_other_groups: list[int] = draw(
+        lists(integers(min_value=0, max_value=1000), min_size=0, max_size=10)
+    )
+    inputs: list[bytes] = draw(lists(task_inputs(), min_size=1, max_size=10))
+    identifier_pointers: list[int] = draw(
+        lists(integers(min_value=0, max_value=1000), min_size=len(inputs), max_size=len(inputs))
+    )
+    cluster_slug_pointers: list[int] = draw(
+        lists(integers(min_value=0, max_value=1000), min_size=len(inputs), max_size=len(inputs))
+    )
+    display_pointers: list[int] = draw(
+        lists(integers(min_value=0, max_value=1000), min_size=len(inputs), max_size=len(inputs))
+    )
+    max_retries_values: list[int] = draw(
+        lists(integers(min_value=0, max_value=100), min_size=len(inputs), max_size=len(inputs))
+    )
+
+    return TaskSubmissionGroup(
+        dependencies_on_other_groups,
+        inputs,
+        identifier_pointers,
+        cluster_slug_pointers,
+        display_pointers,
+        max_retries_values,
+    )
+
+
+@composite
+def task_submissions(draw: DrawFn) -> TaskSubmissions:
+    """A hypothesis strategy for generating random task_submissions"""
+    task_groups = draw(lists(task_submission_groups(), min_size=1, max_size=10))
+    cluster_slug_lookup = draw(lists(alphanumerical_text(), min_size=1, max_size=10))
+    identifier_lookup = draw(lists(task_identifiers(), min_size=1, max_size=10))
+    display_lookup = draw(lists(alphanumerical_text(), min_size=1, max_size=10))
+
+    return TaskSubmissions(task_groups, cluster_slug_lookup, identifier_lookup, display_lookup)
+
+
+@composite
+def single_task_submissions(draw: DrawFn) -> SingleTaskSubmission:
     """A hypothesis strategy for generating random sub_tasks"""
     cluster_slug = str(draw(uuids(version=4)))
     identifier = draw(task_identifiers())
-    task_input = draw(task_inputs())
-    dependencies: list[int] = []
+    inputs = draw(task_inputs())
+    dependencies: list[int] = draw(lists(integers(min_value=0, max_value=1000), min_size=0, max_size=10))
     display = draw(alphanumerical_text())
     max_retries = draw(integers(min_value=0, max_value=100))
 
-    return TaskSubmission(cluster_slug, identifier, task_input, dependencies, display, max_retries)
+    return SingleTaskSubmission(cluster_slug, identifier, inputs, dependencies, display, max_retries)
 
 
 @composite
@@ -166,7 +248,7 @@ def computed_tasks(draw: DrawFn) -> ComputedTask:
     """A hypothesis strategy for generating random computed_tasks"""
     task_id = draw(uuids(version=4))
     display = draw(alphanumerical_text())
-    subtasks: list[TaskSubmission] = draw(lists(task_submissions(), min_size=1, max_size=10))
+    subtasks = draw(task_submissions())
     progress_updates = draw(lists(progress_indicators(), min_size=0, max_size=3))
 
     return ComputedTask(task_id, display, subtasks, progress_updates)
@@ -215,7 +297,7 @@ def automations(draw: DrawFn) -> AutomationPrototype:
     """A hypothesis strategy for generating random automations"""
     automation_id = draw(uuids(version=4))
     name = draw(alphanumerical_text())
-    prototype = draw(task_submissions())
+    prototype = draw(single_task_submissions())
 
     storage_event = []
     cron = []
@@ -227,3 +309,44 @@ def automations(draw: DrawFn) -> AutomationPrototype:
         storage_event = draw(lists(storage_event_triggers(), min_size=1, max_size=10))
 
     return AutomationPrototype(automation_id, name, prototype, storage_event, cron)
+
+
+@composite
+def time_intervals(draw: DrawFn) -> TimeInterval:
+    start = draw(datetimes())
+    end = draw(datetimes())
+    start, end = min(start, end), max(start, end)  # make sure start is before end
+    start_exclusive = draw(booleans())
+    end_inclusive = draw(booleans())
+    return TimeInterval(start, end, start_exclusive, end_inclusive)
+
+
+@composite
+def id_intervals(draw: DrawFn) -> IDInterval:
+    start = draw(uuids(version=4))
+    end = draw(uuids(version=4))
+    start, end = min(start, end), max(start, end)  # make sure start is before end
+    start_exclusive = draw(booleans())
+    end_inclusive = draw(booleans())
+    return IDInterval(start, end, start_exclusive, end_inclusive)
+
+
+@composite
+def query_filters(draw: DrawFn) -> QueryFilters:
+    """A hypothesis strategy for generating random query filters"""
+    use_time_interval = draw(booleans())
+    time_interval = None
+    id_interval = None
+    if use_time_interval:
+        time_interval = draw(time_intervals())
+    else:
+        id_interval = draw(id_intervals())
+
+    automation_ids = draw(lists(uuids(version=4), min_size=0, max_size=10))
+
+    job_states = draw(lists(sampled_from(JobState), min_size=0, max_size=10))
+    if job_states is not None:
+        job_states = list(set(job_states))  # de-duplicate
+
+    name = draw(alphanumerical_text() | none())
+    return QueryFilters(time_interval, id_interval, automation_ids, job_states, name)